Fix instance directions for fractional predicates
The `as_fractional_fractional` coercion is problematic - it goes in a
weird direction when doing an instance search for `Fractional`, and it
loops with the naturally-defined `fractional_as_fractional` instance
included in this commit.

Generally, we only want to search from a known proposition to a
potential fractional split, rather than the other way around (indeed,
this is the only way the search goes almost anywhere in the codebase.)
The coercion, and its associated hint mode, don't make a lot of sense,
since they go against this idea.

This commit removes the coercion and its associated hint mode. This
change broke a few proofs, which incidentally relied on the hint mode to
make progress when using some other typeclass instances with
`AsFractional` bounds. To fix these proofs, the relevant instances have
been adjusted - in every case, the breakage was due to the instances not
respecting the correct input-output class argument roles, and trying to
apply searches from outputs to inputs: the hint mode broke these
searches because it prevented progress on the outputs in an
`AsFractional` bound.